+++ packages/e2e-tests/tests/optimisticRecordTracking.ts
@@ -1,6 +1,7 @@
 import {
   DepositManager,
   SimpleERC20Token__factory,
+  Teller,
 } from "@nocturne-xyz/contracts";
 import { SimpleERC20Token } from "@nocturne-xyz/contracts/dist/src/SimpleERC20Token";
 import {
@@ -28,6 +29,7 @@ describe("Optimistic nullifier tracking", () => {
 
   let sdk: NocturneWalletSDK;
   let db: NocturneDB;
+  let teller: Teller;
   let depositManager: DepositManager;
   let eoa: ethers.Wallet;
 
@@ -50,7 +52,7 @@ describe("Optimistic nullifier tracking", () => {
       },
     });
 
-    ({ teardown, fillSubtreeBatch, depositManager } = testDeployment);
+    ({ teardown, fillSubtreeBatch, depositManager, teller } = testDeployment);
 
     eoa = testDeployment.aliceEoa;
 
@@ -97,7 +99,10 @@ describe("Optimistic nullifier tracking", () => {
 
     // make op request spending 200 tokens
     const amountToSpend = 200n;
-    const opRequest = new OperationRequestBuilder()
+    const opRequest = new OperationRequestBuilder({
+      chainId: 31337n,
+      tellerContract: teller.address,
+    })
       .unwrap(erc20Asset, amountToSpend)
       .action(erc20.address, encodedTransfer)
       .gasPrice(0n)
@@ -105,7 +110,6 @@ describe("Optimistic nullifier tracking", () => {
         BigInt((await depositManager.provider.getBlock("latest")).timestamp) +
           ONE_DAY_SECONDS
       )
-      .chainId(31337n)
       .build();
 
     // prepare op
@@ -210,7 +214,10 @@ describe("Optimistic nullifier tracking", () => {
 
     // make op request spending 200 tokens
     const amountToSpend = 200n;
-    const opRequest = new OperationRequestBuilder()
+    const opRequest = new OperationRequestBuilder({
+      chainId: 31337n,
+      tellerContract: teller.address,
+    })
       .unwrap(erc20Asset, amountToSpend)
       .action(erc20.address, encodedTransfer)
       .gasPrice(0n)
@@ -218,7 +225,6 @@ describe("Optimistic nullifier tracking", () => {
         BigInt((await depositManager.provider.getBlock("latest")).timestamp) +
           ONE_DAY_SECONDS
       )
-      .chainId(31337n)
       .build();
 
     // prepare op