+++ packages/sdk/src/opRequestGas.ts
@@ -419,6 +419,7 @@ function fakeProvenOperation(op: Operation): ProvenOperation {
     };
   });
   return {
+    networkInfo: op.networkInfo,
     refundAddr: op.refundAddr,
     encodedRefundAssets: op.encodedRefundAssets,
     actions: op.actions,
@@ -428,7 +429,6 @@ function fakeProvenOperation(op: Operation): ProvenOperation {
     maxNumRefunds: op.maxNumRefunds,
     gasPrice: op.gasPrice,
     joinSplits: provenJoinSplits,
-    chainId: op.chainId,
     deadline: op.deadline,
     atomicActions: op.atomicActions,
   };